requirements.txt
colorama==0.4.6
openai==0.27.2
python-dotenv==1.0.0
